requirements.txt
appdirs==1.4.4
attrs==20.3.0
beautifulsoup4==4.9.3
black==19.10b0
click==7.1.2
et-xmlfile==1.0.1
jdcal==1.4.1
loguru==0.5.3
mypy-extensions==0.4.3
openpyxl==3.0.6
pathspec==0.8.1
regex==2020.11.13
selenium==3.141.0
soupsieve==2.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
appdirs==1.4.4
attrs==20.3.0
beautifulsoup4==4.9.3
black==19.10b0
click==7.1.2
et-xmlfile==1.0.1
jdcal==1.4.1
loguru==0.5.3
mypy-extensions==0.4.3
openpyxl==3.0.7
pathspec==0.8.1
python-dotenv==0.15.0
regex==2020.11.13
selenium==3.141.0
soupsieve==2.2
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
